move buffer pool manager callbacks from src to test
