+++ src/CMakeLists.txt
@@ -11,4 +11,3 @@ file(GLOB_RECURSE murmur3_sources
         ${PROJECT_SOURCE_DIR}/third_party/murmur3/*.cpp ${PROJECT_SOURCE_DIR}/third_party/murmur3/*.h)
 add_library(thirdparty_murmur3 SHARED ${murmur3_sources})
 target_link_libraries(bustub_shared thirdparty_murmur3)
-target_link_libraries(bustub_shared gtest)
