+++ src/include/recovery/log_record.h
@@ -105,7 +105,7 @@ class LogRecord {
         prev_lsn_(prev_lsn),
         log_record_type_(log_record_type),
         prev_page_id_(prev_page_id) {
-    // calculate log record size
+    // calculate log record size, header size + sizeof(prev_page_id) + sizeof(page_id)
     size_ = HEADER_SIZE + sizeof(page_id_t) * 2;
   }
